+++ assembler/src/hammer/hamcluster.cpp
@@ -59,7 +59,7 @@ std::pair<size_t, size_t> SubKMerSplitter::split() {
 }
 
 #if 1
-static bool canMerge(const ConcurrentDSU &uf, int x, int y) {
+static bool canMerge(const ConcurrentDSU &uf, unsigned x, unsigned y) {
   size_t szx = uf.set_size(x), szy = uf.set_size(y);
   const size_t hardthr = 2500;
 
@@ -76,7 +76,7 @@ static bool canMerge(const ConcurrentDSU &uf, int x, int y) {
   return true;
 }
 #else
-static bool canMerge(const ConcurrentDSU &uf, int x, int y) {
+static bool canMerge(const ConcurrentDSU &uf, unsigned x, unsigned y) {
   return (uf.set_size(x) + uf.set_size(y)) < 10000;
 }
 #endif
@@ -88,10 +88,10 @@ static void processBlockQuadratic(ConcurrentDSU  &uf,
                                   unsigned tau) {
   size_t blockSize = block.size();
   for (size_t i = 0; i < blockSize; ++i) {
-    size_t x = block[i];
+    unsigned x = (unsigned)block[i];
     hammer::KMer kmerx = data[x].kmer();
-    for (uint32_t j = i + 1; j < blockSize; j++) {
-      size_t y = block[j];
+    for (size_t j = i + 1; j < blockSize; j++) {
+      unsigned y = (unsigned)block[j];
       hammer::KMer kmery = data[y].kmer();
       if (uf.find_set(x) != uf.find_set(y) &&
           canMerge(uf, x, y) &&
